remove console logs
