+++ pages/clothes-condition/recycle.js
@@ -3,7 +3,7 @@ import Breadcrumb from "../../components/Breadcrumb";
 import dynamic from "next/dynamic";
 import Search from "../../components/SearchRecycle";
 import UpdateCount from "../../components/UpdateCount";
-
+import SearchByUserLocation from "../../components/SearchByUserLocation";
 //this was needed to get the full map to load, rather than just a couple of squares
 //I don't fully understand how it's working, but setting server-side rendering to false means that the map is dynamically loaded on the client side
 const LondonMap = dynamic(() => import("../../components/Map"), { ssr: false });
@@ -25,7 +25,6 @@ export default function Recycle({ userPosition, setUserPosition }) {
   }, [userInput])
 
   async function handleSearch() {
-    console.log(userInput);
     const result = await fetch(
       `../api/recylePoint?abc=${userInput.toUpperCase()}
       `
@@ -56,46 +55,7 @@ export default function Recycle({ userPosition, setUserPosition }) {
 
 
 
-  async function handleSearchFromLocation(pos) {
-    console.log(pos);
-    const crd = pos.coords;
 
-    console.log('Your current position is:');
-    console.log(`Latitude : ${crd.latitude}`);
-    console.log(`Longitude: ${crd.longitude}`);
-
-
-    let lat = crd.latitude;
-
-    let lon = crd.longitude;
-
-    // let lon = 0.629834723775309
-    // let lat = 51.7923246977375
-
-    const postcode = await fetch(`https://api.postcodes.io/postcodes?lon=${lon}&lat=${lat}`);
-    const postJson = await postcode.json();
-    let userPostcode = postJson.result[0].postcode;
-    await setUserInput(userPostcode);
-
-    // await handleSearch();
-
-    if (!postcode.ok) {
-      setData(null);
-      setError(
-        `Oops, looks like we can't locate you. Enter a postcode in the search bar.`
-      );
-      return;
-    }
-  }
-
-  function geoError(err) {
-    console.warn(`ERROR(${err.code}): ${err.message}`);
-  }
-
-  function locator() {
-    console.log('here');
-    navigator.geolocation.getCurrentPosition(handleSearchFromLocation, geoError);
-  }
 
 
 
@@ -112,13 +72,11 @@ export default function Recycle({ userPosition, setUserPosition }) {
         handleSearch={handleSearch}
         labelText={"Enter your postcode..."}
       />
-      <form
-        onSubmit={(e) => {
-          e.preventDefault();
-          // fetchData(postcode);
-        }}>
-        <button onClick={() => { locator() }}>Find Me</button>
-      </form>
+      <SearchByUserLocation
+        setUserInput={setUserInput}
+        setData={setData}
+        setError={setError}
+      />
       {
         data ? (
           <button
